mcp-servers/requirements.txt
fastmcp==2.13.0.2
pydantic==2.12.3
python-dotenv==1.2.1

python-multipart>=0.0.9,<1
openpyxl
pandas>=2.2,<3
scikit-learn>=1.4,<2